+++ packages/expo-linear-gradient/CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### 🎉 New features
 
+- Mark React client components with "use client" directives.
 - Added support for Apple tvOS. ([#26965](https://github.com/expo/expo/pull/26965) by [@douglowder](https://github.com/douglowder))
 
 ### 🐛 Bug fixes